fix: rename dto
